application-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
bcrypt==4.1.2
brave-search==0.2.1
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
deptry==0.23.0
dnspython==2.7.0
fake-http-header==0.3.5
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
geopy==2.4.1
google==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.21.1
google-generativeai>=0.2.0
Google_Search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
huggingface-hub==0.32.2
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langfuse==3.0.3
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.4
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.5.0
mypy==1.10.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.2
passlib[bcrypt]==1.7.4
pdfplumber==0.10.4
pillow==11.2.1
playwright==1.52.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pymongo==4.13.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
requirements-parser==0.13.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.4
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
sympy==1.14.0
tenacity==8.2.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.4.0
uv==0.7.13
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0